+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Activities.DetailActivity" android:noHistory="true" android:theme="@style/Theme.AppCompat.Translucent" />
+        <activity android:name=".Activities.OverviewActivity" android:noHistory="true" android:theme="@style/Theme.AppCompat.Translucent" />
     </application>
 
 </manifest>
