+++ app/src/main/res/values-v21/styles.xml
@@ -16,7 +16,7 @@
 
         <style name="Theme.AppCompat.Translucent">
                 <item name="android:windowNoTitle">true</item>
-                <item name="android:windowBackground">@android:color/transparent</item>
+                <item name="android:windowBackground">@color/colorGrayTransparent</item>
                 <item name="android:colorBackgroundCacheHint">@null</item>
                 <item name="android:windowIsTranslucent">true</item>
                 <item name="android:windowAnimationStyle">@android:style/Animation</item>
