+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,5 @@
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
     <color name="colorTransparent">#00FFFFFF</color>
+    <color name="colorGrayTransparent">#66000000</color>
 </resources>
